WORKING
in the current state of the gmae, you can operate it like a tycoon game. there are some issues that need to be resolved such as selling of buildings and upgrades and that money going into the account of the user. however there are "overdraft protectoins" that prevent the user from going into debt. the main thing to fix now are the loans which need to be added and to add in the framework for instructions. it would also be nice to be able to save progress in the game. the list of improvemnts and changes also ofcorse inclueds adding a GUI to the game for the user to interact with